Comentários do swagger

+++ Backend/WSTowerApi/Controllers/UsuarioController.cs
@@ -24,28 +24,90 @@ namespace WSTowerApi.Controllers
             _usuarioRepository = new UsuarioRepository();
         }
 
+        /// <summary>
+        /// Cadastra um novo usuario
+        /// </summary>
+        /// <param name="novoUsuario">Objeto com as informações</param>
+        /// <returns>Um status code 201 - Created</returns>
+        /// <response code="201">Retorna apenas o status code Created</response>
+        /// <response code="400">Retorna o erro gerado</response>
+        /// dominio/api/Usuarios
         [HttpPost]
         public IActionResult Post(Usuario novoUsuario)
         {
-            _usuarioRepository.Cadastrar(novoUsuario);
+            try
+            {
+                _usuarioRepository.Cadastrar(novoUsuario);
 
-            return StatusCode(201);
+                return StatusCode(201);
+            }
+            catch (Exception error)
+            {
+                return BadRequest(error);
+            }
         }
 
+        /// <summary>
+        /// Atualiza um usuário existente
+        /// </summary>
+        /// <param name="id">ID do usuário que será atualizado</param>
+        /// <param name="usuarioAtualizado">Objeto com as novas informações</param>
+        /// <returns>Um status code 204 - No Content</returns>
+        /// <response code="204">Retorna apenas o status code No Content</response>
+        /// <response code="404">Retorna uma mensagem de erro</response>
+        /// <response code="400">Retorna o erro gerado</response>
+        /// dominio/api/Usuarios/id
         [HttpPut]
         public IActionResult Put(int id, Usuario usuarioAtualizado)
         {
-            _usuarioRepository.Atualizar(id, usuarioAtualizado);
+            try
+            {
+                Usuario usuarioBuscado = _usuarioRepository.ListarPorId(id);
 
-            return StatusCode(204);
+                if (usuarioBuscado != null)
+                {
+                    _usuarioRepository.Atualizar(id, usuarioAtualizado);
+
+                    return StatusCode(204);
+                }
+
+                return NotFound("Nenhum usuário encontrado para o ID informado");
+            }
+            catch (Exception error)
+            {
+                return BadRequest(error);
+            }
         }
 
+        /// <summary>
+        /// Deleta um usuário
+        /// </summary>
+        /// <param name="id">ID do usuário que será deletado</param>
+        /// <returns>Um status code 202 - Accepted</returns>
+        /// <response code="202">Retorna apenas o status code Accepted</response>
+        /// <response code="404">Retorna uma mensagem de erro</response>
+        /// <response code="400">Retorna o erro gerado</response>
+        /// dominio/api/Usuarios/id
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
-            _usuarioRepository.Remover(id);
+            try
+            {
+                Usuario usuarioBuscado = _usuarioRepository.ListarPorId(id);
 
-            return StatusCode(204);
+                if (usuarioBuscado != null)
+                {
+                    _usuarioRepository.Remover(id);
+
+                    return StatusCode(202);
+                }
+
+                return NotFound("Nenhum usuário encontrado para o ID informado");
+            }
+            catch (Exception error)
+            {
+                return BadRequest(error);
+            }
         }
     }
 }
